fix manage trust repo

+++ cli.sh
@@ -747,9 +747,10 @@ git hooks trust [forget] [--local|--global]
         if [ "$1" = "revoke" ]; then
             return
         fi
+    
     fi
 
-    if [ "$1" = "delete" ] || [ -f .githooks/trust-all ]; then
+    if [ "$1" = "delete" ] && [ -f .githooks/trust-all ]; then
         rm -rf .githooks/trust-all &&
             echo "The trust marker is removed from the repository." &&
             echo_if_non_bare_repo "  Do not forget to commit and push the change!" &&
@@ -757,6 +758,7 @@ git hooks trust [forget] [--local|--global]
 
         echo "! Failed to delete the trust marker" >&2
         exit 1
+
     fi
 
     echo "! Unknown subcommand: $1" >&2
